+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Added checks for missing ACKs from the I2C slave.
+  Missing ACKs will now return an `NakError` from the I2C traits.
+
 ## [0.5.1] - 2021-03-20
 ### Fixed
 - Fixed the I2C `Read` trait not setting the read address bit.